+++ src/GUI/gui.py
@@ -35,6 +35,7 @@ from MainMenu import recent
 from MainMenu import theme_changer
 from GUI import snippet_code
 from API import kilo 
+from GUI import right_panel
 
 class MainWindow(ct.CTk):
     def __init__(self, *args, **kwargs):
@@ -65,12 +66,8 @@ class MainWindow(ct.CTk):
         file_drop.add_option(option="New File                     Ctrl+N", command=lambda: file_menu.custom_file(statusbar_instance, treeview_frame))
         file_drop.add_option(option="Open                          Ctrl+O",command=lambda:file_menu.open_file(scroll.text, scroll, statusbar_instance))
         file_drop.add_separator()
-        file_drop.add_option(option="Open Input",command=lambda:file_menu.open_input(treeview_frame))
-        file_drop.add_option(option="Save Input",command=lambda:file_menu.save_input(treeview_frame))
-        file_drop.add_option(option="Open Output",command=lambda:file_menu.open_output(treeview_frame))
-        file_drop.add_separator()
-        file_drop.add_option(option="Open Folder               Ctrl+K",command=lambda:file_menu.open_folder(treeview_frame, statusbar_instance, scroll))
-        file_drop.add_option(option="Close Folder", command=lambda:file_menu.close_folder(treeview_frame, scroll))
+        file_drop.add_option(option="Open Folder               Ctrl+K",command=lambda:file_menu.open_folder(treeview_frame, statusbar_instance, scroll, right_panel_frame))
+        file_drop.add_option(option="Close Folder", command=lambda:file_menu.close_folder(treeview_frame, scroll, right_panel_frame))
         file_drop.add_option(option="Open Recent", command=lambda:open_recent(self))
         file_drop.add_separator()
         file_drop.add_option(option="Save                           Ctrl+S",command=lambda:file_menu.save_file(scroll.text, statusbar_instance))
@@ -105,8 +102,8 @@ class MainWindow(ct.CTk):
         view_drop.add_separator()
         view_drop.add_option(option="Status Bar",command=lambda:view_menu.hide_unhide_statusbar(statusbar_instance))
         view_drop.add_option(option="Notifications",command=lambda:view_menu.notifications(statusbar_instance))
-        view_drop.add_option(option="Left Panel                     Ctrl+B", command=lambda:view_menu.hide_unhide_treeview(treeview_frame, scroll))
-        view_drop.add_option(option="Input & Output", command=lambda:view_menu.hide_unhide_input_output(treeview_frame))
+        view_drop.add_option(option="Left Panel                     Ctrl+B", command=lambda:view_menu.hide_unhide_treeview(treeview_frame, scroll, right_panel_frame))
+        view_drop.add_option(option="Right Panel             Ctrl+Alt+B", command=lambda:view_menu.hide_unhide_right_panel(scroll, right_panel_frame, treeview_frame))
         view_drop.add_option(option="Terminal                         Ctrl+`", command=lambda:scroll.handle_terminal())
         view_drop.add_separator()
         view_drop.add_option(option="Refresh editor     Ctrl+Shift+R", command=lambda:scroll.redraw())
@@ -151,7 +148,8 @@ class MainWindow(ct.CTk):
                                                                                          scroll.tab_bar))
         
         utility_drop = CustomDropdownMenu(widget=utility, font=("", 20), corner_radius=4, separator_color="#b0b0b0")
-        utility_drop.add_option(option="Run                                F5", command=lambda:run.run_cpp_file(treeview_frame, scroll.text))
+        utility_drop.add_option(option="Run                                F5", command=lambda:run.run_cpp_file(scroll.text))
+        utility_drop.add_option(option="Run with pre-input", command=lambda:run.pre_input_run(self))
         utility_drop.add_option(option="Paint Mode              Ctrl+P", command=lambda:open_paint_mode(self))
         utility_drop.add_separator()
         utility_drop.add_option(option="Start Server (Beta)", command= lambda:scroll.start_server())
@@ -167,6 +165,9 @@ class MainWindow(ct.CTk):
         self.scroll = scroll
         treeview_frame = treeview.TreeviewFrame(self, scroll, statusbar_instance, scroll)
         
+        right_panel_frame = right_panel.RightPanel(self)
+        self.right_panel_frame = right_panel_frame
+
         self.auto_save_option()
 
         self.grid_columnconfigure(0, weight=1)
@@ -179,7 +180,7 @@ class MainWindow(ct.CTk):
 
         scroll.text.focus()
         self.after(200, scroll.redraw())
-        statusbar_instance.run_img.bind("<Button-1>", lambda event: run.run_cpp_file(treeview_frame, scroll.text))
+        statusbar_instance.run_img.bind("<Button-1>", lambda event: run.run_cpp_file(scroll.text))
         scroll.text.bind("<Control-n>", lambda event: file_menu.custom_file(statusbar_instance, treeview_frame))
         scroll.text.bind("<Control-o>", lambda event: file_menu.open_file(scroll.text, scroll, statusbar_instance))
         scroll.text.bind("<Control-s>", lambda event: file_menu.save_file(scroll.text, statusbar_instance))
@@ -193,12 +194,12 @@ class MainWindow(ct.CTk):
         scroll.text.bind("<Control-f>", lambda event: edit_menu.find_text(scroll.text, scroll))
         scroll.text.bind("<Control-h>", lambda event:edit_menu.replace_text(scroll.text, scroll))
         scroll.text.bind("<F11>", lambda event: view_menu.toggle_fullscreen(self))
-        scroll.text.bind("<Control-k>",lambda event:file_menu.open_folder(treeview_frame, statusbar_instance, scroll))
-        scroll.text.bind("<Control-b>", lambda event: view_menu.hide_unhide_treeview(treeview_frame, scroll))
+        scroll.text.bind("<Control-k>",lambda event:file_menu.open_folder(treeview_frame, statusbar_instance, scroll, right_panel_frame))
+        scroll.text.bind("<Control-b>", lambda event: view_menu.hide_unhide_treeview(treeview_frame, scroll, right_panel_frame))
+        scroll.text.bind("<Control-Alt-b>", lambda event:view_menu.hide_unhide_right_panel(scroll, right_panel_frame, treeview_frame))
         scroll.text.bind("<Control-Shift-T>", lambda event: template_menu.use_template(scroll.text, scroll, statusbar_instance))
-        scroll.text.bind("<F5>", lambda event:run.run_cpp_file(treeview_frame, scroll.text))
+        scroll.text.bind("<F5>", lambda event:run.run_cpp_file(scroll.text))
         scroll.text.bind("<Control-p>", lambda event:open_paint_mode(self))
-        treeview_frame.input.bind("<Control-s>", lambda event:file_menu.save_input(treeview_frame))
         scroll.text.bind("<Control-Shift-R>", lambda event:scroll.redraw())
         scroll.text.bind("<Control-MouseWheel>", lambda event:self.mouse_wheel(event=event))
         pywinstyles.apply_dnd(scroll.text, self.drag_file)
@@ -213,7 +214,7 @@ class MainWindow(ct.CTk):
             file_menu.open_default_file(scroll.text, scroll, statusbar_instance)
 
         if not check.get_config_value("default_folder") == "0":
-            file_menu.open_folder(treeview_frame, statusbar_instance, scroll, check.get_config_value("default_folder"))
+            file_menu.open_folder(treeview_frame, statusbar_instance, scroll,right_panel_frame, check.get_config_value("default_folder"))
 
         if not check.get_config_value("files") == "0":
             session.load_file_tab(scroll.tab_bar)