+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 16. Kilonova.ro tools 
 17. Python support
 18. Scrollbar in suggestions
+19. Right panel
+20. New input & output system
 
 ### Fixed/Changes: